feat: refactor user service
